refactor: relocate API middleware
