+++ app/api/team/invites/accept/route.ts
@@ -3,13 +3,9 @@ import { z } from 'zod';
 import { prisma } from '@/lib/database/prisma';
 import { getServerSession } from 'next-auth';
 import { authOptions } from '@/lib/auth/index';
-import {
-  createSuccessResponse,
-  withErrorHandling,
-  withValidation,
-  ApiError,
-  ERROR_CODES
-} from '@/lib/api/common';
+import { createSuccessResponse, ApiError, ERROR_CODES } from '@/lib/api/common';
+import { withErrorHandling } from '@/middleware/error-handling';
+import { withValidation } from '@/middleware/validation';
 import { createTeamMemberNotFoundError } from '@/lib/api/team/error-handler';
 
 const acceptInviteSchema = z.object({ token: z.string() });
