+++ src/lib/api/common/index.ts
@@ -2,10 +2,8 @@
  * API Common Utilities
  * 
  * This file exports all common utilities for API routes.
- * It provides a single import point for error handling, response formatting, and middleware.
  */
 
 export * from './error-codes';
 export * from './api-error';
 export * from './response-formatter';
-export * from './middleware';
